Relocate retrieve stream log tests
